More unit tests for CRAN skeleton
