+++ tests/test-cran-skeleton/rpart/DESCRIPTION
@@ -1,0 +1,32 @@
+Package: rpart
+Priority: recommended
+Version: 4.1-15
+Date: 2019-04-10
+Authors@R: c(person("Terry", "Therneau", role = "aut",
+	            email = "therneau@mayo.edu"),
+             person("Beth", "Atkinson", role = c("aut", "cre"),
+	            email = "atkinson@mayo.edu"),
+             person("Brian", "Ripley", role = "trl",
+                    email = "ripley@stats.ox.ac.uk",
+		    comment = "producer of the initial R port, maintainer 1999-2017"))
+Description: Recursive partitioning for classification, 
+  regression and survival trees.  An implementation of most of the 
+  functionality of the 1984 book by Breiman, Friedman, Olshen and Stone.
+Title: Recursive Partitioning and Regression Trees
+Depends: R (>= 2.15.0), graphics, stats, grDevices
+Suggests: survival
+License: GPL-2 | GPL-3
+LazyData: yes
+ByteCompile: yes
+NeedsCompilation: yes
+Author: Terry Therneau [aut],
+  Beth Atkinson [aut, cre],
+  Brian Ripley [trl] (producer of the initial R port, maintainer
+    1999-2017)
+Maintainer: Beth Atkinson <atkinson@mayo.edu>
+Repository: CRAN
+URL: https://github.com/bethatkinson/rpart,
+        https://cran.r-project.org/package=rpart
+BugReports: https://github.com/bethatkinson/rpart/issues
+Packaged: 2019-04-11 15:59:38 UTC; atkinson
+Date/Publication: 2019-04-12 14:32:39 UTC
